+++ django_athm/ath_movil/conf/defaults.py
@@ -1,4 +1,0 @@
-DJANGO_ATHM_SANDBOX_MODE = False
-ATHM_PUBLIC_TOKEN = None
-ATHM_PRIVATE_TOKEN = None
-ATHM_CALLBACK_VIEW = "django_athm.ath_movil.views.callback"