+++ tests/test_utils.py
@@ -1,6 +1,0 @@
-# from django_athm.ath_movil import utils
-
-
-class TestHTTPAdapter:
-    def test(self):
-        pass